feat(page): edit tab UI overhaul + sidebar design language
Kekule composer drawing fixes:
- composer.newDoc() after init (upstream commented out its own newDoc
  call, so the editor starts with no ChemObj and silently drops clicks)
- Skip RightSidebar's pointerdown stopper inside .kekule-container so
  Kekule widgets (which listen at document level) receive events
- Drop CSS scale(0.75) hack; Kekule mixes offsetLeft/Top with clientX/Y
  so any transform breaks pointer->canvas math
- Theme icons via filter on .K-Pri-Glyph-Content (Kekule uses CSS
  background-image, not img)

Kekule container auto-sizes: setDimension 100%/100% + composer's own
ResizeObserver; remove outer ScrollArea; add contentClassName to
SidebarSection so sections can flex-fill.

New DownloadStructureSection: PubChem source (name/CID -> IsomericSMILES
-> existing handleSmiles pipeline); other sources stubbed with "(soon)".

Sidebar UI Design Language documented in CLAUDE.md (type scale, control
sizing, spacing, colour semantics, writing style). Align Measure /
Manipulate / Select / View panels to the language.

Fix core test suite & hook failures (molrs 0.2 f32->f64 rename):
- Tests pass Float64Array to setColF / Box.cube
- probeColumnDtype returns "f64"; update dtype checks across
  entity_source / modifiers / data_inspector / reader / app
- Grid.getArray returns Float64Array directly (no .toCopy wrapper);
  widen marchingCubes input signature
- Convert MolvisSessionRegistry from static-only class to module
  (noStaticOnlyClass / noThisInStatic)
- Use optional chaining in jsonrpc colorBy guard

+++ core/src/analysis/rdf.ts
@@ -20,11 +20,11 @@ export interface RdfParams {
 
 export interface RdfResult {
   /** Bin center distances */
-  r: Float32Array;
+  r: Float64Array;
   /** g(r) values */
-  gr: Float32Array;
+  gr: Float64Array;
   /** Raw pair counts per bin */
-  counts: Float32Array;
+  counts: Float64Array;
   /** Number of bins */
   nBins: number;
   /** Bin width */
@@ -187,9 +187,9 @@ function buildSubFrame(frame: Frame, indices: number[]): Frame | null {
   if (!x || !y || !z) return null;
 
   const n = indices.length;
-  const sx = new Float32Array(n);
-  const sy = new Float32Array(n);
-  const sz = new Float32Array(n);
+  const sx = new Float64Array(n);
+  const sy = new Float64Array(n);
+  const sz = new Float64Array(n);
   for (let i = 0; i < n; i++) {
     sx[i] = x[indices[i]];
     sy[i] = y[indices[i]];
